+++ source/ProcessManager.Core/Infrastructure/Scheduling/RecurringOrchestrationQueries.cs
@@ -1,0 +1,40 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.ProcessManagement.Core.Application.Scheduling;
+using Energinet.DataHub.ProcessManagement.Core.Domain.OrchestrationDescription;
+using Energinet.DataHub.ProcessManagement.Core.Infrastructure.Database;
+using Microsoft.EntityFrameworkCore;
+
+namespace Energinet.DataHub.ProcessManagement.Core.Infrastructure.Scheduling;
+
+/// <summary>
+/// Queries necessary to handle the planning of recurring orchestrations.
+/// </summary>
+internal class RecurringOrchestrationQueries(
+        ProcessManagerContext context) :
+            IRecurringOrchestrationQueries
+{
+    private readonly ProcessManagerContext _context = context;
+
+    /// <inheritdoc />
+    public async Task<IReadOnlyCollection<OrchestrationDescription>> GetAllRecurringAsync()
+    {
+        var query = _context.OrchestrationDescriptions
+            .Where(x => x.IsEnabled == true)
+            .Where(x => x.RecurringCronExpression != string.Empty);
+
+        return await query.ToListAsync().ConfigureAwait(false);
+    }
+}
